backend/pyproject.toml
[tool.poetry]
name = "minerva-backend"
version = "0.1.0"
description = "Backend processing engine for Minerva"
authors = ["Alex <alexander.elgier@gmail.com>"]
readme = "README.md"
packages = [{ include = "minerva_backend", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.112"
uvicorn = "^0.30"
neo4j = "^5.23"
ollama = "^0.1.0"
minerva-models = {path = "../../minerva-models", develop = true}
aiosqlite = "^0.21.0"
temporalio = {extras = ["testing"], version = "^1.17.0"}
pytest-asyncio = "^1.2.0"
dependency-injector = "^4.41.0"
pydantic-settings = "^2.0.0"
intervaltree = "^3.1.0"
rapidfuzz = "^3.14.1"
diskcache = "^5.6.3"
requests = "^2.32.5"
pandas = "^2.3.3"
ipykernel = "^6.30.1"
pip = "^25.2"
jupyterlab = "^4.4.9"
matplotlib = "^3.10.6"
langchain = "^1.0.0"
langgraph = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.0"
black = "^24.0"
isort = "^6.0"
flake8 = "^7.0"
radon = "^6.0"
vulture = "^2.10"
mypy = "^1.8"
sphinx = "^7.0"
mkdocs = "^1.5"
mkdocs-material = "^9.4"
langgraph-cli = {extras = ["inmem"], version = "^0.4.4"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Code Quality Tools Configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["minerva_backend"]
known_third_party = ["fastapi", "neo4j", "temporalio", "pydantic", "pytest"]

[tool.mypy]
python_version = "3.13"
warn_return_any = false
warn_unused_configs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
warn_unreachable = false
strict_equality = false
ignore_missing_imports = true
disable_error_code = ["attr-defined", "misc", "override", "import-untyped"]

[tool.radon]
exclude = ["tests/", "docs/", "build/", "dist/"]
min = "B"
show_complexity = true
show_duplicates = true